pyproject.toml
[tool.poetry]
name = "kafka"
version = "0.1.0"
description = ""
authors = ["Rhanyele Marinho <rhany.marinho@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "==3.12.2"
confluent-kafka = "^2.3.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
